Replaces serializer test with previous version's
The old version of ember-template had a blank serializer test.
Issue #23 is open because of this, but isn't something we should
be resolving now. The focus of this merge is to de-podify, which
I am hopefully completing by moving the serializer test to the right
directory, and deleting the (now obsolete) tests from the pod-structure.